requirements.txt
appdirs==1.4.4
asgiref==3.2.7
aubio==0.4.9
audiofile==0.2.2
audioread==2.1.8
autopep8==1.5.5
cffi==1.14.0
cfgv==3.2.0
decorator==4.4.2
distlib==0.3.1
Django==3.0.7
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
filelock==3.0.12
identify==1.5.10
joblib==0.15.1
llvmlite==0.32.1
Markdown==3.2.1
mysqlclient==1.4.6
nodeenv==1.5.0
numba==0.49.1
numpy==1.18.4
pre-commit==2.7.1
pycodestyle==2.6.0
pycparser==2.20
pytz==2019.3
PyYAML==5.3.1
resampy==0.2.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
SoundFile==0.10.3.post1
sox==1.3.7
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.2
virtualenv==20.2.2
Wave==0.0.2
